Remove single quotes from destination paths that include spaces.
When 'toPath' contains one or more spaces, it is wrapped in single quotes.  If 'toPath' also begins with a volume identifier (e.g. 'C:'), the opening single quote offsets the colon (:) by one position in the string.  This results in a 'NotSupportedException' exception being thrown by the call to 'GetFullPath()'.

The solution is to always trim single quotes from 'toPath' before the call to 'GetFullPath()' is made.